Reorganize code, recalculate when we save campaign, adapttests
