+++ plugins/polio/tests/test_api.py
@@ -686,7 +686,7 @@ class PreparednessAPITestCase(APITestCase):
         self.client = APIClient()
         self.client.force_authenticate(self.yoda)
 
-    def test_two_campaign_round(self):
+    def test_two_campaign_round_empty(self):
         campaign_a = Campaign.objects.create(obr_name="campaign A", account=self.account)
         round_one = campaign_a.rounds.create(number=1)
         round_three = campaign_a.rounds.create(number=3)
@@ -699,7 +699,20 @@ class PreparednessAPITestCase(APITestCase):
         self.assertEqual(len(r["rounds"]), 2)
 
         response = self.client.get(f"/api/polio/preparedness_dashboard/", format="json")
-        self.assertJSONResponse(response, 200)
+        r = self.assertJSONResponse(response, 200)
+        self.assertEqual(len(r), 0)
+
+    def test_two_campaign_round_error(self):
+        campaign_a = Campaign.objects.create(obr_name="campaign A", account=self.account)
+        round_one = campaign_a.rounds.create(number=1)
+        round_three = campaign_a.rounds.create(number=3)
+        Campaign.objects.create(obr_name="campaign B", account=self.account)
+        Campaign.objects.create(obr_name="campaign c", account=self.account)
+
+        response = self.client.get(f"/api/polio/campaigns/{campaign_a.id}/", format="json")
+
+        r = self.assertJSONResponse(response, 200)
+        self.assertEqual(len(r["rounds"]), 2)
 
         round_one.preparedness_spreadsheet_url = "https://docs.google.com/spreadsheets/d/1"
         round_one.save()
@@ -711,7 +724,13 @@ class PreparednessAPITestCase(APITestCase):
         self.assertEqual(len(r), 2)
         for campaign_round in r:
             self.assertEqual(campaign_round["status"], "not_sync")
-        print(r)
+
+        """[{'campaign_id': '6ecda204-2206-4ae2-a38a-5359684dccf6', 'campaign_obr_name': 'campaign A',
+          'indicators': {}, 'round': 'Round1', 'round_id': 44, 'round_start': None, 'round_end': None,
+          'status': 'not_sync', 'details': 'This spreadsheet has not been synchronised yet'}, {
+                'campaign_id': '6ecda204-2206-4ae2-a38a-5359684dccf6', 'campaign_obr_name': 'campaign A',
+                'indicators': {}, 'round': 'Round3', 'round_id': 45, 'round_start': None, 'round_end': None,
+                'status': 'not_sync', 'details': 'This spreadsheet has not been synchronised yet'}]"""
 
 
 class TeamAPITestCase(APITestCase):
